עדכון

+++ BLL/Convertion/CitiesConvertion.cs
@@ -31,7 +31,7 @@ namespace BLL.Convertion
             {
                 Id = city.Id,
                 Name = city.Name,
-                AreaId = city.AreaId
+                AreaId = (int)city.AreaId
             };
         }
         public static List<city> Convert(List<CityDTO> city)
